Lock screen fix: Make sure user app can write memory/pwrmgr/state
Now that the power manager is running under the user-session (as user
"app"), it was unable to modify the memory/pwrmgr/state key due to
insufficient permissions.

By permitting the "app" user to write this key, the lock screen is
able to launch before the device enters the sleep state, and it is
functional when the device wakes up.

Change-Id: I4519d949c5186c0437abedba9bb7469799fedcde
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>